settings in own file
