+++ templates/roots/tracking_tab.html
@@ -78,7 +78,7 @@
 
 
 {# TODO: put somewhere else #}
-<svg style="position:absolute" id="svg-defs">
+<svg style="position:absolute; pointer-events:none;" id="svg-defs">
 <defs>
   <marker id="dot-marker" viewBox="0 0 4 4" refX="2" refY="2" markerHeight="2" markerWidth="2">
     <circle r="2" cx="2" cy="2" fill="blue"></circle>
